+++ libs/rxjs/array/src/index.ts
@@ -6,7 +6,12 @@
 
 export { binarySearch } from './lib/binary-search';
 export { difference } from './lib/difference';
+export { differenceWith } from './lib/difference-with';
 export { intersects } from './lib/intersects';
+export { intersectsWith } from './lib/intersects-with';
+export { sort } from './lib/sort';
+export { sortMap } from './lib/sort-map';
+export { flipArray } from './lib/flip-array';
 
 export * from './types/array-compare';
 export * from './types/binary-search';