+++ libs/rxjs/array/src/lib/sort.spec.ts
@@ -1,5 +1,3 @@
-import { of } from 'rxjs';
-import { intersectsWith } from './intersects-with';
 import { marbles } from 'rxjs-marbles';
 import { sort } from './sort';